requirements.txt
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
plotly==5.18.0
scikit-learn==1.2.2
dash==2.14.2
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
scipy==1.11.4
lmfit==1.2.0
openpyxl==3.1.2
openai==1.3.7
python-dotenv==1.0.0
werkzeug==2.3.7
flask==2.3.3 